+++ src/Parsers/TimeUnit.cs
@@ -1,0 +1,11 @@
+﻿namespace Espeon {
+    public enum TimeUnit {
+        SECOND = 1,
+        MINUTE = 60,
+        HOUR = 3600,
+        DAY = 86400,
+        WEEK = 604800,
+        MONTH = 2628000,
+        YEAR = 31536000
+    }
+}
